refactor(fast-element): reorganize internals
