+++ packages/web-components/fast-element/src/components/attributes.ts
@@ -1,6 +1,6 @@
-import { Accessor, Observable } from "./observation/observable";
-import { DOM } from "./dom";
-import { Notifier } from "./observation/notifier";
+import { Accessor, Observable } from "../observation/observable";
+import { DOM } from "../dom";
+import { Notifier } from "../observation/notifier";
 
 /**
  * Represents objects that can convert values to and from
