+++ packages/web-components/fast-element/src/index.ts
@@ -1,31 +1,34 @@
 export * from "./platform";
-export * from "./template";
-export * from "./fast-element";
-export { FASTElementDefinition, PartialFASTElementDefinition } from "./fast-definitions";
-export * from "./attributes";
-export * from "./controller";
-export * from "./interfaces";
-export * from "./template-compiler";
+export * from "./templating/template";
+export * from "./components/fast-element";
 export {
-    css,
-    CSSDirective,
+    FASTElementDefinition,
+    PartialFASTElementDefinition,
+} from "./components/fast-definitions";
+export * from "./components/attributes";
+export * from "./components/controller";
+export * from "./interfaces";
+export * from "./templating/compiler";
+export {
     ElementStyles,
     ElementStyleFactory,
     ComposableStyles,
     StyleTarget,
-} from "./styles";
-export * from "./view";
+} from "./styles/element-styles";
+export { css } from "./styles/css";
+export { CSSDirective } from "./styles/css-directive";
+export * from "./templating/view";
 export * from "./observation/observable";
 export * from "./observation/notifier";
 export { Splice } from "./observation/array-change-records";
 export { enableArrayObservation } from "./observation/array-observer";
 export { DOM } from "./dom";
-export * from "./directives/behavior";
-export * from "./directives/binding";
-export * from "./directives/directive";
-export * from "./directives/ref";
-export * from "./directives/when";
-export * from "./directives/repeat";
-export * from "./directives/slotted";
-export * from "./directives/children";
-export { elements, NodeBehaviorOptions } from "./directives/node-observation";
+export * from "./observation/behavior";
+export * from "./templating/binding";
+export * from "./templating/html-directive";
+export * from "./templating/ref";
+export * from "./templating/when";
+export * from "./templating/repeat";
+export * from "./templating/slotted";
+export * from "./templating/children";
+export { elements, NodeBehaviorOptions } from "./templating/node-observation";
