+++ packages/web-components/fast-element/src/styles/element-styles.ts
@@ -1,5 +1,5 @@
-import { Behavior } from "./directives/behavior";
-import { DOM } from "./dom";
+import { Behavior } from "../observation/behavior";
+import { DOM } from "../dom";
 
 /**
  * A node that can be targeted by styles.
@@ -251,83 +251,4 @@ export class StyleElementStyles extends ElementStyles {
     private normalizeTarget(target: StyleTarget): StyleTarget {
         return target === document ? document.body : target;
     }
-}
-
-/**
- * Directive for use in {@link css}.
- *
- * @public
- */
-export class CSSDirective {
-    /**
-     * Creates a CSS fragment to interpolate into the CSS document.
-     * @returns - the string to interpolate into CSS
-     */
-    public createCSS(): ComposableStyles {
-        return "";
-    }
-
-    /**
-     * Creates a behavior to bind to the host element.
-     * @returns - the behavior to bind to the host element, or undefined.
-     */
-    public createBehavior(): Behavior | undefined {
-        return undefined;
-    }
-}
-
-/**
- * Transforms a template literal string into styles.
- * @param strings - The string fragments that are interpolated with the values.
- * @param values - The values that are interpolated with the string fragments.
- * @remarks
- * The css helper supports interpolation of strings and ElementStyle instances.
- * @public
- */
-export function css(
-    strings: TemplateStringsArray,
-    ...values: (ComposableStyles | CSSDirective)[]
-): ElementStyles {
-    const styles: ComposableStyles[] = [];
-    let cssString = "";
-    const behaviors: Behavior[] = [];
-
-    for (let i = 0, ii = strings.length - 1; i < ii; ++i) {
-        cssString += strings[i];
-        let value = values[i];
-
-        if (value instanceof CSSDirective) {
-            const behavior = value.createBehavior();
-            value = value.createCSS();
-
-            if (behavior) {
-                behaviors.push(behavior);
-            }
-        }
-
-        if (value instanceof ElementStyles || value instanceof CSSStyleSheet) {
-            if (cssString.trim() !== "") {
-                styles.push(cssString);
-                cssString = "";
-            }
-
-            styles.push(value);
-        } else {
-            cssString += value;
-        }
-    }
-
-    cssString += strings[strings.length - 1];
-
-    if (cssString.trim() !== "") {
-        styles.push(cssString);
-    }
-
-    const elementStyles = ElementStyles.create(styles);
-
-    if (behaviors.length) {
-        elementStyles.withBehaviors(...behaviors);
-    }
-
-    return elementStyles;
 }
