+++ packages/web-components/fast-element/src/templating/children.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 import { children, ChildrenBehavior } from "./children";
-import { AttachedBehaviorHTMLDirective } from "./directive";
+import { AttachedBehaviorHTMLDirective } from "./html-directive";
 import { observable } from "../observation/observable";
 import { elements } from "./node-observation";
 import { DOM } from "../dom";
