+++ packages/web-components/fast-element/src/templating/when.ts
@@ -1,4 +1,4 @@
-import { CaptureType, SyntheticViewTemplate } from "../template";
+import { CaptureType, SyntheticViewTemplate } from "./template";
 import { ExecutionContext, Binding } from "../observation/observable";
 
 /**
